+++ src/Services/Flow.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace OVAC\Guardrails\Services;
+
+class Flow extends FlowExtensionBuilder
+{
+}
+