+++ src/Support/helpers.php
@@ -1,0 +1,9 @@
+<?php
+
+if (!function_exists('flow')) {
+    function flow(): \OVAC\Guardrails\Services\Flow
+    {
+        return new \OVAC\Guardrails\Services\Flow();
+    }
+}
+